move filtering and bbox
